Move _LoadedModelTracker to pyfunc
